+++ template/base_feature/src/main/java/ru/surfstudio/standard/application/app/App.kt
@@ -3,6 +3,7 @@ package ru.surfstudio.standard.application.app
 import android.app.Application
 import android.content.Context
 import android.content.Intent
+import android.os.StrictMode
 import com.akaita.java.rxjava2debug.RxJava2Debug
 import com.github.anrwatchdog.ANRWatchDog
 import com.google.firebase.crashlytics.FirebaseCrashlytics
@@ -15,6 +16,7 @@ import ru.surfstudio.android.notification.ui.PushEventListener
 import ru.surfstudio.android.template.base_feature.BuildConfig
 import ru.surfstudio.android.template.base_feature.R
 import ru.surfstudio.android.utilktx.ktx.ui.activity.ActivityLifecycleListener
+import ru.surfstudio.android.utilktx.util.SdkUtils
 import ru.surfstudio.standard.application.app.di.AppInjector
 import ru.surfstudio.standard.application.logger.strategies.remote.FirebaseCrashlyticsRemoteLoggingStrategy
 import ru.surfstudio.standard.application.logger.strategies.remote.RemoteLoggerLoggingStrategy
@@ -31,6 +33,7 @@ class App : Application() {
 
         initAnrWatchDog()
         initLog()
+        initVmPolicy()
 
         RxJavaPlugins.setErrorHandler { Logger.e(it) }
         AppInjector.initInjector(this)
@@ -63,6 +66,16 @@ class App : Application() {
         Logger.addLoggingStrategy(TimberLoggingStrategy())
         Logger.addLoggingStrategy(RemoteLoggerLoggingStrategy())
         RemoteLogger.addRemoteLoggingStrategy(FirebaseCrashlyticsRemoteLoggingStrategy())
+    }
+
+    private fun initVmPolicy() {
+        if (SdkUtils.isAtLeastS()) {
+            val policy = StrictMode.VmPolicy.Builder()
+                .detectUnsafeIntentLaunch()
+                .build()
+
+            StrictMode.setVmPolicy(policy)
+        }
     }
 
     private fun initRxJava2Debug() {