+++ template/f-onboarding/src/main/java/ru/surfstudio/standard/f_onboarding/di/OnboardingScreenConfigurator.kt
@@ -1,6 +1,6 @@
 package ru.surfstudio.standard.f_onboarding.di
 
-import android.os.Bundle
+import android.content.Intent
 import dagger.Component
 import dagger.Module
 import dagger.Provides
@@ -13,55 +13,54 @@ import ru.surfstudio.android.core.mvp.configurator.ScreenComponent
 import ru.surfstudio.android.dagger.scope.PerScreen
 import ru.surfstudio.standard.f_onboarding.*
 import ru.surfstudio.standard.ui.activity.di.ActivityComponent
-import ru.surfstudio.standard.ui.activity.di.FragmentScreenConfigurator
-import ru.surfstudio.standard.ui.navigation.routes.OnboardingFragmentRoute
+import ru.surfstudio.standard.ui.activity.di.ActivityScreenConfigurator
+import ru.surfstudio.standard.ui.navigation.routes.OnboardingActivityRoute
 import ru.surfstudio.standard.ui.screen_modules.ActivityScreenModule
 import ru.surfstudio.standard.ui.screen_modules.CustomScreenModule
-import ru.surfstudio.standard.ui.screen_modules.FragmentScreenModule
 
 /**
- * Конфигуратор [OnboardingFragmentView].
+ * Конфигуратор [OnboardingActivityView].
  */
-internal class OnboardingScreenConfigurator(bundle: Bundle?) : FragmentScreenConfigurator(bundle) {
+internal class OnboardingScreenConfigurator(intent: Intent) : ActivityScreenConfigurator(intent) {
 
     @PerScreen
     @Component(
-            dependencies = [ActivityComponent::class],
-            modules = [FragmentScreenModule::class, OnboardingScreenModule::class]
+        dependencies = [ActivityComponent::class],
+        modules = [ActivityScreenModule::class, OnboardingScreenModule::class]
     )
-    internal interface OnboardingScreenComponent : BindableScreenComponent<OnboardingFragmentView>
+    internal interface OnboardingScreenComponent : BindableScreenComponent<OnboardingActivityView>
 
     @Module
-    internal class OnboardingScreenModule(route: OnboardingFragmentRoute) :
-            CustomScreenModule<OnboardingFragmentRoute>(route) {
+    internal class OnboardingScreenModule(route: OnboardingActivityRoute) :
+        CustomScreenModule<OnboardingActivityRoute>(route) {
 
         @Provides
         @PerScreen
         fun provideEventHub(screenEventHubDependency: ScreenEventHubDependency) =
-                ScreenEventHub<OnboardingEvent>(screenEventHubDependency, OnboardingEvent::Lifecycle)
+            ScreenEventHub<OnboardingEvent>(screenEventHubDependency, OnboardingEvent::Lifecycle)
 
         @Provides
         @PerScreen
         fun provideBinder(
-                screenBinderDependency: ScreenBinderDependency,
-                eventHub: ScreenEventHub<OnboardingEvent>,
-                middleware: OnboardingMiddleware,
-                reducer: OnboardingReducer,
-                stateHolder: OnboardingStateHolder
+            screenBinderDependency: ScreenBinderDependency,
+            eventHub: ScreenEventHub<OnboardingEvent>,
+            middleware: OnboardingMiddleware,
+            reducer: OnboardingReducer,
+            stateHolder: OnboardingStateHolder
         ): Any = ScreenBinder(screenBinderDependency).apply {
             bind(eventHub, middleware, stateHolder, reducer)
         }
     }
 
     override fun createScreenComponent(
-            activityComponent: ActivityComponent?,
-            fragmentScreenModule: FragmentScreenModule?,
-            args: Bundle?
+        activityComponent: ActivityComponent,
+        activityScreenModule: ActivityScreenModule,
+        intent: Intent
     ): ScreenComponent<*> {
         return DaggerOnboardingScreenConfigurator_OnboardingScreenComponent.builder()
-                .activityComponent(activityComponent)
-                .fragmentScreenModule(fragmentScreenModule)
-                .onboardingScreenModule(OnboardingScreenModule(OnboardingFragmentRoute()))
-                .build()
+            .activityComponent(activityComponent)
+            .activityScreenModule(activityScreenModule)
+            .onboardingScreenModule(OnboardingScreenModule(OnboardingActivityRoute()))
+            .build()
     }
 }